+++ internal/schemautil/userconfig/const.go
@@ -1,0 +1,9 @@
+package userconfig
+
+const (
+	// SchemaPackage is the fully-qualified package name of the schema package.
+	SchemaPackage = "github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+
+	// SchemaUtilPackage is the fully-qualified package name of the schemautil package.
+	SchemaUtilPackage = "github.com/aiven/terraform-provider-aiven/internal/schemautil"
+)